Changes to interpolation implementation with a new 'dummy' file option available for single analysis results.
Decided on a linear interpolation of the log-likelihood curves.  The
2nd/3rd order spline interpolations would miss the 0,0 point and added
‘noise’ to the median null TS measurement, hurting the sensitivity.
The spline interpolation code is not yet removed and can be
re-activated.

Also, sensitivity with a single analysis result is now possible.  A
trivial no-effect ‘merge’ can be performed by specifying ‘dummy’ as the
second input file name.  This maximizes the log-likelihood of just one
analysis in the standard format and can be piped into the sensitivity
script.