Added two utility scripts to shuffle and examine files
Added shuffle.py script to reorder the results of any set of scrambled
trials.  This is a type of resampling that is useful for checking that
the order in which we combined experimental trials is not special.  The
script also groups trials at the same flux level together in the output.

The ntrials.py script is useful in case you just want to check how many
trials were run at each flux level in a mysterious file.

Reordered main test_data file.